#4567 split what's common and what's different between HU moving processes
https://github.com/metasfresh/metasfresh/issues/4567
